telegram-bot/requirements.txt
# constraints
--constraint=constraints.txt

# requirements
python-telegram-bot
requests
